sparklin/BlobTriggerFuncApp/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-data-tables
azure-functions==1.7.2
azure-identity==1.6.1
pyapacheatlas==0.12
